backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
pymongo>=4.7
pandas==2.2.2
openpyxl==3.1.4
gTTS==2.5.1

